+++ rust-toolchain.toml
@@ -2,5 +2,5 @@
 # cargo + rustup will use this to consistently build the project
 # with the same version across all checkouts and environments
 [toolchain]
-channel = "1.79.0"
+channel = "ms-1.82.0"
 profile = "default"
